Add tests to check new encryption code
Also renamed existing SQLCipher tests to make distinction clear.
